refactor(escalation): promote tools to subpackage with shared seam
Quick-form escalation was a near-verbatim copy of the app-task
escalation factory — same recipient resolution, same task-title
preamble, same result post-processing, same LangGraph wrapper, with
only the platform task-creation call genuinely differing.  Reorganise
into agent/tools/escalation/ (mirroring agent/tools/mcp/) and lift the
shared scaffolding into a common.py seam so the per-variant factories
own only what is actually variant-specific.

* common.py — EscalationAction, recipient/asset resolution, output
  schema builder, invocation preamble, post-interrupt finaliser, and
  the LangGraph tool wrapper factory.
* app_task.py, quick_form.py, ixp_vs.py — assemble those primitives;
  escalation-memory persistence stays inside app_task.py (its only
  caller).
* memory.py — escalation_memory.py moved under the subpackage.
