+++ src/uipath_langchain/agent/tools/__init__.py
@@ -2,13 +2,15 @@
 
 from .a2a import A2aClient, create_a2a_tools_and_clients, open_a2a_tools
 from .context_tool import create_context_tool
-from .escalation_tool import create_escalation_tool
+from .escalation import (
+    create_escalation_tool,
+    create_ixp_escalation_tool,
+    create_quick_form_escalation_tool,
+)
 from .extraction_tool import create_ixp_extraction_tool
 from .integration_tool import create_integration_tool
-from .ixp_escalation_tool import create_ixp_escalation_tool
 from .mcp import open_mcp_tools
 from .process_tool import create_process_tool
-from .quick_form_escalation_tool import create_quick_form_escalation_tool
 from .tool_factory import (
     create_tools_from_resources,
 )
