+++ src/uipath_langchain/agent/tools/escalation/__init__.py
@@ -1,0 +1,33 @@
+"""Action Center escalation tools.
+
+Three escalation variants share a single resource concept
+(``AgentResourceType.ESCALATION``) but differ in how the HITL task is
+materialised:
+
+* :func:`create_escalation_tool` — ``escalationType=0``, Action Center
+  app task (with optional escalation memory).
+* :func:`create_ixp_escalation_tool` — ``escalationType=1``, Document
+  Understanding validation action.
+* :func:`create_quick_form_escalation_tool` — ``escalationType=2``,
+  schema-first FormLib task.
+
+All three are assembled from shared primitives in :mod:`.common`.
+"""
+
+from .app_task import create_escalation_tool
+from .common import (
+    EscalationAction,
+    resolve_asset,
+    resolve_recipient_value,
+)
+from .ixp_vs import create_ixp_escalation_tool
+from .quick_form import create_quick_form_escalation_tool
+
+__all__ = [
+    "EscalationAction",
+    "create_escalation_tool",
+    "create_ixp_escalation_tool",
+    "create_quick_form_escalation_tool",
+    "resolve_asset",
+    "resolve_recipient_value",
+]
