+++ tests/agent/guardrails/actions/test_escalate_action.py
@@ -249,7 +249,7 @@ class TestEscalateAction:
     )
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_sends_correct_data(
         self,
         mock_resolve_recipient,
@@ -327,7 +327,7 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_post_agent_with_agent_result(
         self,
         mock_resolve_recipient,
@@ -401,7 +401,7 @@ class TestEscalateAction:
             GuardrailScope.TOOL,
         ],
     )
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_post_execution_single_message_raises_error(
         self, mock_resolve_recipient, scope: GuardrailScope
     ):
@@ -434,7 +434,7 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_tool_pre_execution_extracts_tool_args(
         self,
         mock_resolve_recipient,
@@ -494,7 +494,7 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_tool_post_execution_extracts_tool_content(
         self,
         mock_resolve_recipient,
@@ -556,7 +556,7 @@ class TestEscalateAction:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_post_execution_ai_message_with_tool_calls(
         self,
         mock_resolve_recipient,
@@ -1726,7 +1726,7 @@ class TestEscalateAction:
     )
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_resolves_recipient_correctly(
         self,
         mock_resolve_recipient,
@@ -1768,7 +1768,7 @@ class TestEscalateAction:
         assert call_kwargs["recipient"] == expected_value
 
     @pytest.mark.asyncio
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_with_asset_recipient_resolution_failure(
         self, mock_resolve_recipient
     ) -> None:
@@ -1846,7 +1846,7 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_standard_recipient_assigned_to_uses_value(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -1888,7 +1888,7 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_standard_recipient_assigned_to_uses_display_name(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -1930,7 +1930,7 @@ class TestEscalateActionMetadata:
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.interrupt")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_asset_recipient_assigned_to_uses_resolved_value(
         self, mock_resolve_recipient, mock_interrupt, mock_uipath_class, mock_config
     ):
@@ -1972,7 +1972,7 @@ class TestEscalateActionMetadata:
     @pytest.mark.asyncio
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPathConfig")
     @patch("uipath_langchain.agent.guardrails.actions.escalate_action.UiPath")
-    @patch("uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value")
+    @patch("uipath_langchain.agent.tools.escalation.resolve_recipient_value")
     async def test_create_task_node_sets_metadata_node_type(
         self, mock_resolve_recipient, mock_uipath_class, mock_config
     ):
