+++ tests/cli/test_agent_with_guardrails.py
@@ -1168,7 +1168,7 @@ graph = create_agent(
                         return_value=mock_uipath_instance,
                     ),
                     patch(
-                        "uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value",
+                        "uipath_langchain.agent.tools.escalation.resolve_recipient_value",
                         mock_resolve,
                     ),
                 ):
@@ -1378,7 +1378,7 @@ graph = create_agent(
                         return_value=mock_uipath_instance,
                     ),
                     patch(
-                        "uipath_langchain.agent.tools.escalation_tool.resolve_recipient_value",
+                        "uipath_langchain.agent.tools.escalation.resolve_recipient_value",
                         mock_resolve,
                     ),
                 ):
